Moved to cleaner AutoMake/AutoConf setup. Added BGL::Path::alignTo() call, and fixed a couple geometry bugs. Centralized SVG file header generation. Made several BGL tests visually clearer. Changed extraneous int32_t types to int.

+++ src/BGL/BGLLine.h
@@ -29,7 +29,7 @@ public:
     // Member variables
     Point startPt;
     Point endPt;
-    int16_t flags;
+    int flags;
     double temperature;
     double extrusionWidth;
